Add #rebalance method, update #insert and #remove
Added the #rebalance method that rebalances the
binary tree.

Updated the #insert and #remove methods so that
the recursive logic goes to the #insert_recursive
and #remove_recursive and after the insert or remove
is done then the tree is checked if it is balanced.
If it isn't it is rebalanced.